Add SQL version of highest lowest cumulative average
This implements the same calculation as `lowest_cumulative.py`, but in
SQL (specifically, Postgres) rather than Python + pandas.

I personally find this clearer; it's a shell script with some
administrative set up (start Postgres, create a database, and undo those
at the end) and a single SQL script. That script just creates a table,
imports the relevant CSV into it, and builds up the results using a
succession of CTEs:

1. Calculate the cumulative runs and outs for a player's career by
   innings.
2. Use that to calculate the cumulative average for all innings where
   there is an average (i.e. the player has been out at least once).
3. Take the total runs the player scored and the lowest cumulative
   average.
4. From the final CTE, pick the players with the highest lowest
   cumulative average who meet the runs cut-off and display them.

On my machine, I also see this as quite a bit faster than the pandas
version.

    $ time gender=men format=odi ./lowest_cumulative.sh
    # ...
    2.525 secs

    $ time python lowest_cumulative.py -f odi
    # ...
    9.246 secs