Calculate game and team Bannermen
The Bannerman is the percentage of team runs made in a completed
innings. It's named for Charles Bannerman, who made 165 out of 245 (67%)
in the first Test innings ever, which hasn't been beaten.

This script calculate one of two alternatives:

1. The game Bannerman (`type=game ./game_bannerman.sh`). This is for
   games where all 40 wickets fell, and lists the ten players with the
   highest proportion of match runs.
2. The team Bannerman (`type=team ./game_bannerman.sh`). The same as
   above, but only for the team, where the team lost 20 wickets.

Unfortunately there are a couple of insoluble issues based on the data I
have:

1. This only counts runs off the bat. For instance, Bannerman's first
   innings would be counted as 165 / 237 (which is 70%), because there
   were 8 extras.
2. It only counts innings where all the players were dismissed. Again
   using Bannerman's first innings as an example: he retired hurt and
   Tom Garrett was not out. This meant that the team was all out, even
   though there were two batters who end up with not-outs in the
   scorebook.

In both cases, this data is using player-level innings data, and so we
don't know team-level information such as whether the team was all out
or how many extras there were. If we had the type of dismissal or
retirement, we could potentially use that: if only one player is not out
in an innings, and all 10 others were dismissed or retired hurt, we
could say that the team was all out. But we don't, so we can't.

It's also possible to look at the records for women's Tests with
`gender=women`. The highest proportions there are lower, probably simply
because there have been far fewer women's Tests.